* tf_util.c (tf_get_cred): Issue date is written out as a long, 		read back in as same.
I have kept the size as a long to be compatible with the Cygnus V4
distribution. The problem was introduced when we changed the include/kerberosIV
structures to use 32 bit ints for timestamps. So, under OSF/1, tf_util would
write out a 64 bit issue date, and then try reading back in a 32 bit one. Since
Cygnus uses sizeof(long) we will too.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@7667 dc483132-0cff-0310-8789-dd5450dbe970