* Make shared libs work on AIX * DO not use #ifdef unix * move krb5/krb/crypto_glue.c to crypto/krb5_glue.c
	crypto_glue.c contained a bunch of functions that used to be macros.  They do belong in libkrb5, but if they go there,
they create a circular dependency because libcrypto uses them.
The obvious long-term right answer is to get libcrypto to stop using interfaces
with names like krb5_*.
I didn't feel like tackling that in the middle of getting things working on AIX
so I just moved the functions to be declared in the same library as use.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@9949 dc483132-0cff-0310-8789-dd5450dbe970