ticket: new subject: implement krb5_read_password, des_read_password via krb5_prompter_posix
This commit fixes one incompatibility introduced when
krb5_read_password was reimplemented in terms of krb5_prompter_posix.
There is a remaining incompatibility, which is krb5_prompter_posix's
appending of the string ": " following a prompt.  Callers of
krb5_read_password and of des_read_password don't expect this
behavior, which results in a double colon prompt, which breaks the
libkadm5 test suite.

	* read_pwd.c (krb5_read_password): Restore name of size_return.
	Set *size_return after successful call to krb5_prompter_posix,
	since some callers were actually checking, e.g. kadm5.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@14922 dc483132-0cff-0310-8789-dd5450dbe970