Ticket: 881 Status: open
Between the 1.1 and 1.2 releases of Kerberos, the gss-sample protocol
changed.  People seem to use this sample for interop testing, so
support for the old protocol is needed.

This commit adds a -v1 flag to the gss-client that will use the old
protocol.

The server does not yet support the old protocol, so  the ticket remains open.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15103 dc483132-0cff-0310-8789-dd5450dbe970