ticket: 677
Remove ADD_DEF and uses thereof, and move the appropriate symbols to
AC_DEFINE'ed things or to Makefile.in files.  Remove explicit settings
of CPPFLAGS.  These allow CPPFLAGS to be a user parameter.  Also, add
CFLAGS to the CC_LINK variables, so they can control usage of the
compiler to link programs.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15214 dc483132-0cff-0310-8789-dd5450dbe970