ticket: new Subject: GSS_C_NO_CREDENTIAL should accept any principal Tags: enhancement cc: nicolas.williams@sun.com cc: krbdev@mit.edu
If a context is accepted with GSS_C_NO_CREDENTIAL or if a credential
is acquired with GSS_C_NO_NAME as the acceptor name  then allow any
principal in the keytab  to be  used as the acceptor name.

This means that gss_inquire_cred can return GSS_C_NO_NAME from a
credential.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15218 dc483132-0cff-0310-8789-dd5450dbe970