Ticket: new Target_Version: 1.3 Subject: Register writable keytabs by default Component: krb5-libs Tags: pullup
We have a customer requirement to support writable keytabs using the
public API in 1.3 Discussion on krbcore indicates there is no good
reason why these are not registered by default.  So, they are now
registered by default.

Also adjust other code in the tree not to try and register them.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15459 dc483132-0cff-0310-8789-dd5450dbe970