ticket:new subject: Memory leaks in ccache due to thread integration tags: pullup
	* cc_file.c (krb5_fcc_close): Free the cache id.
	(dereference): When removing fcc_set entry from list, free the
	pointer as well.

The first was accidently dropped in the dereference code writing.
The cache id pointer is never freed.

The second error is the removal of the krb5_fcc_data from the linked list.
The fcc_set is removed from the chain, but the memory for the removed fcc_set
is never freed.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@16981 dc483132-0cff-0310-8789-dd5450dbe970