ticket: new
Some applications such as Eudora on Windows load and unload the Kerberos
libraries as part of a plug-in.  Plugins are often loaded for a specific
purpose and then unregistered.  In order to support this model, the libraries
must restore the library state to the uninitialized state when the library
is unloaded.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17050 dc483132-0cff-0310-8789-dd5450dbe970