ticket: new subject: Tru64 compilation fails after k5-int.h/krb5.h changes target_version: 1.4.4 status: open
Due to some silliness in db-config.h and the Tru64 system header files, an
accidental change in the order of inclusion of certain headers caused the build
to fail.

With this patch, "make all" succeeds, but "make check" fails partway through;
I'm still investigating, and don't know at this point if there are additional
compilation problems.

* policy_db.h: Include db.h after gssrpc/types.h, to fix compilation on Tru64.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17515 dc483132-0cff-0310-8789-dd5450dbe970