ticket: 3253
Remove the case sensitive comparisions of registry entry
and schema strings.  Microsoft uses HKLM\"SOFTWARE" and
HKCU\"Software".  This means the encoding schema that was
selected does not work and the conflict in case prevents
plugins from being loaded.   

Better to enable plugins to work than to allow two realms
that differ only by case on the same platform during the 
beta.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17523 dc483132-0cff-0310-8789-dd5450dbe970