ticket: new subject: Wix MSI installer for KFW 3.0 Beta 3 component: windows status: resolved target_version: 1.4.4 tags: pullup
Add operating system version check to ensure it is not
installed on Windows 95, 98, ME or NT 4.0

Break out the license text into its own file

Restructure the installer to allow either NetIdMgr or Leash
to be selected (by transform only) as the credentials manager.
The default is to use NetIdMgr.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17533 dc483132-0cff-0310-8789-dd5450dbe970