ticket: new subject: Network Identity Manager updates for KFW 3.0 Beta 3 component: windows status: open target_version: 1.4.4
Fix the handling of case sensitive names being stored in the 
registry.  Only apply case sensitive encoding logic to the
keys below the NetIdMgr key.

Fix the importing of credentials from MSLSA:

Apply an ugly hack to krb5configcc.c that forces _WIN32_WINNT
to 0x0501 for the one file so that the executable can be built
as APPVER=5.0 and yet still gain access to balloon tips on XP
and above.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17535 dc483132-0cff-0310-8789-dd5450dbe970