ticket: 3416 status: open
Remove the thread-safety flag from the kdb plugin interface.  Instead,
have the kdb code assume the plugin is thread safe, and implement some
quick and dirty wrapper functions in the db2 plugin to make it use a
local mutex.

There's still some mutex code in the kdb library that should be
reviewed, and simplified or removed.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17611 dc483132-0cff-0310-8789-dd5450dbe970