ticket: 3415 tags: pullup
Qing Dong <dongq@mit.edu> provided a set of changes to allow
krb5 to build under the Microsoft Visual Studio 8 compiler 
in 64-bit mode and produce file names that do not conflict 
with the names produced by the 32-bit build.  That patch 
was modified to work on Unix and also include processor
dependent pre-processor definitions to remove warnings.




git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17680 dc483132-0cff-0310-8789-dd5450dbe970