ticket: 3542 status: open
identity/plugins/common/dynimport.c:
During the interop session we concluded that the ccapi32.dll should
not be required for netidmgr to operate.  netidmgr should work with
only FILE: ccaches.  After the interop the removal of the error
check post-load was not removed.  

identity/doc/Makefile:
The 'clean' rules failed to specify the /Q switch which silently
removes the directory tree.  As a result, during the build the
user was prompted.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17907 dc483132-0cff-0310-8789-dd5450dbe970