ticket: new subject: krb5 1.5 alpha - memory leaks in krb5kdc due to not freeing error messages tags: pullup
In the kdc and lib/kadm5/logger.c, krb5_get_error_message needs to be paired
with krb5_free_error_message to release returned memory.  

Essentially a memory leak was introduced for every principal requested
that did not exist in the database.

Identified by valgrind on the kdc - running kdc_hammer and specifying
more principals than are present in the db.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18072 dc483132-0cff-0310-8789-dd5450dbe970