ticket: 3887
Remove -lkrb5support, which should be for internal use only, from the displayed
list of libraries.  This means the output is not suitable for linking against
static libraries built from the tree, but prevents us from committing to
keeping libkrb5support.1 in all future Mac shared-library builds (in addition
to the libkrb5support.0 we're stuck with for non-Mac UNIXy shared-library
builds).


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18154 dc483132-0cff-0310-8789-dd5450dbe970