Ticket: 3122
Pass in the correct value for the first argument of select (one larger
than the largest file number in the select set) rather than some multiple
of sizeof some struct.  The latter is large enough accidentally work, but
breaks on AIX.

Map IPv4-mapped IPv6 addresses back to IPv4 in krshd for the purposes of
connecting back to the remote system on AIX, since on AIX getnameinfo
returns such addresses but connect won't accept them.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18166 dc483132-0cff-0310-8789-dd5450dbe970