ticket: new subject: check calling conventions specified for Windows
When we're making changes on UNIX, sometimes we update the Windows
export list but forget to make sure we've annotated the function
declaration in the header file with a calling convention
specification.

This patch checks the krb5 and gssapi public headers against the
Windows export lists (which are annotated with calling-convention and
other info in comments), and flags any inconsistencies in public
interfaces.

* util/def-check.pl: Be quiet about normal stuff by default; accept a
"-v" option to be verbose.  Exit with non-zero status if something
wrong is detected.  Fix some problems in parsing gssapi header files.
Handle DECSCRIPTION and HEAPSIZE directives in .def files, and DATA
annotation in comments.

* include/Makefile.in (verify-calling-conventions-krb5): New target.
(all-unix): Depend on it in maintainer mode.

* lib/gssapi/Makefile.in (verify-calling-conventions-gssapi): New target.
(all-unix): Depend on it in maintainer mode.
(merged-gssapi-header.h): New target; assemble public headers into one input
file.
(clean-misc-unix): New target; delete merged-gssapi-header.h.
(clean-unix): Depend on it.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18258 dc483132-0cff-0310-8789-dd5450dbe970