ticket: new subject: osf1 -oldstyle_liblookup typo
Not really relevant anymore - as we do not support static linking now.
But in ticket 927 (r16776) - a test was added to determine if gnu ld 
was in use and change the linker flags accordingly.  The variable in
aclocal.m4 was krb5_cv_prog_gnu_ld and this was testing for 
krb5_cv_gnu_ld.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18732 dc483132-0cff-0310-8789-dd5450dbe970