ticket: new subject: Delay kadmind random number initialization until after fork Component: krb5-admin Version_Reported: 1.4.4 Target_Version 1.6 Tags: pullup
Delay initialization of the random number generator in kadmind until
after the fork and backgrounding of the process.  Otherwise, a lack of
sufficient entropy during the system boot process will delay system
boot on systems that run each init script in series and that start
kadmind via an init script.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18793 dc483132-0cff-0310-8789-dd5450dbe970