Subject: MEMORY keytab krb5_kt_get_entry sets enctype to 0 if unspecified Ticket: new
If one invokes krb5_kt_get_entry() for a memory keytab w/ an unspecified 
enctype (i.e. 0) - the returned keytab has it's enctype set to 0 as well.

Also - in copying out the found keytab_entry - when a kvno is unspecified, 
the last match found is used  - not the one with the highest vno.
This was caused by copying out the variable "entry" and not "match".



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19141 dc483132-0cff-0310-8789-dd5450dbe970