ticket: new target_version: 1.6.1 tags: pullup subject: valgrind detects uninitialized (but really unused) bytes in 'queue'
The gsstest program exports a GSSAPI security context to a blob in
memory, writes that memory to a file, and reads it back to use it.
Under valgrind, the writing phase triggers a warning about
uninitialized storage.

The "queue" structure as implemented in generic/util_ordering.c holds
an array of values, some of which may never be initialized.  As far as
I can tell, those uninitialized values are never used before being
initialized, either, but valgrind doesn't know that.

This patch zaps the structure contents (including the array) before
using the queue object.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19196 dc483132-0cff-0310-8789-dd5450dbe970