ticket: new subject: osf1: get proper library dependencies installed target_version: 1.6.1 tags: pullup
The LDCOMBINE setting for Tru64 left out $(CFLAGS), which meant the
-pthread option wasn't being passed in, so the libraries didn't
indicate a dependency on the pthread library despite the fact that we
need it unconditionally (if thread support is enabled).  Including
$(CFLAGS) should fix this.

Our own binaries get linked with -pthread anyways, so they build okay,
but other binaries trying to link against or load our libraries could
have problems.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19208 dc483132-0cff-0310-8789-dd5450dbe970