ticket: new subject: Network Identity Manager 1.2 User Manual component: windows tags: pullup
The NIM 1.2 User Manual includes a completely re-written
introduction to what is NIM as well as new text describing
the new default "basic" view mode and the revised "new credentials"
dialog.  Updates to the menu structures, the toolbar, and 
the modifications to the options pages are all provided.

All images have been updated with the new color scheme.
Image shots were taken on XP SP2 with the Silver XP Theme.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19373 dc483132-0cff-0310-8789-dd5450dbe970