ticket: 5532 tags: pullup
In testing it has been observed that remote desktop connections
will execute the NPLogonNotify function but if the logon is 
re-connecting to an existing session, the LogonScript is ignored.
This leaves orphaned credential cache files.  

This commit adds a function, KFW_cleanup_orphaned_caches, which
is called by NPLogonNotify to delete any orphaned cache files.
An orphaned cache file is one that is older than five minutes.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19477 dc483132-0cff-0310-8789-dd5450dbe970