ticket: new subject: kfwlogon corrections for XP component: windows tags: pullup
This patch addresses a problem discovered on some XP systems.
After rundll32.exe starts, the CreateProcess can fail to 
start kfwcpcc.exe if the current directory is not %WinDir%\System32.
CreateProcess() should be called with the lpApplicationName parameter
set to NULL in order to permit the use of the PATH.

Also, in ConfigureLogonScript ensure that the trailing NUL of the
constructed command line is processed when producing the wide 
character version of the string.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19488 dc483132-0cff-0310-8789-dd5450dbe970