Target_Version: 1.6.1 Ticket: 5521 Tags: pullup
Factor repository access out of bkw.pl into repository1.pl.

Modify bkw.pl to use an initial config file to fetch the sources and then use the config file from those sources to do the build.  This way, the description of how to build the sources is in the config file that is part of the sources.  It is possible and probably reasonable for the initial config file to be the same as the tagged version.  Output all the options used.

Add bootstrap.xml - a sample minimal config file, sufficient to fetch the sources from a repository.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19489 dc483132-0cff-0310-8789-dd5450dbe970