Ticket: new
Modify Unix find test to only check the current directory for the non-existent a.tmp.

When the pismere area is under the script area (as in Jeff's setup on afs), the find test slogs through 100MB of files and occasionally finds an a.tmp.  

This correct the erroneous 'unix find not found' errors and improves performance noticeably when using afs.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19509 dc483132-0cff-0310-8789-dd5450dbe970