Ticket: new Subject: Look for unix find command in multiple places Target_Version: 1.6.1 Not tagged yet in case it should not go in 3.2.
Look for the unix find command in a list of directories instead of in only one place.  This way Jeff can install cygwin in c:\tools\cygwin and Ken can have it in C:\cygwin.  The two places in the list are for those two possibilities.  c:\cygwin\bin and c:\tools\cygwin\bin.

Move some prints below where logging starts, so the messages will be in the log.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19534 dc483132-0cff-0310-8789-dd5450dbe970