ticket: new subject: Modify WIX installer to better support upgrading betas component: windows tags: pullup
The WIX installers did not upgrade previous installations with the same
version number as the current package being installed.  This would leave
multiple installations of KFW x.y.z registered as being installed on the
machine even though only the most recent install is being used.

This commit instructs the Windows Installer to uninstall previous installations
with the same version number (the Maximum Upgrade version) as the package
that is being installed while ensuring that the package being installed 
will not be uninstalled if the installation is being modified or repaired.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19541 dc483132-0cff-0310-8789-dd5450dbe970